pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "playlist-analyse-bot"
version = "1.0.0"
description = "Telegram bot for playlist analysis and concert search"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiogram==3.13.1",
    "sqlalchemy==2.0.23",
    "asyncpg==0.29.0",
    "alembic==1.13.1",
    "nest-asyncio==1.6.0",
    "python-dotenv==1.2.1",
    "yandex-music==2.2.0",
    "requests==2.32.5",
    "selenium==4.39.0",
    "google-genai",
]

[project.optional-dependencies]
dev = [
    "pytest==9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov",
]

[tool.pytest.ini_options]
testpaths = ["src/bot/tests", "src/tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/__pycache__/*", "*/venv/*", "*/scripts/*", "*/bot/main.py", "*/main.py", "*/clients/local_concert_client.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

